Split ClassLoader classes into loader package
