+++ src/boss_bus/loader/instantiator.py
@@ -1,33 +1,13 @@
-"""Classes that load dependencies and use them to instantiate classes.
-
-Class loading classes should implement the Interface (ClassLoader)
-"""
+"""Default ClassLoader that instantiates simple classes."""
 
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from typing import Type, TypeVar, get_type_hints, overload
+from typing import Type, get_type_hints, overload
+
+from boss_bus.loader import ClassLoader, obj
 
 RETURN_ANNOTATION = "return"
-
-obj = TypeVar("obj", bound=object)
-
-
-class ClassLoader(ABC):
-    """An Interface that allows loading dependencies and instantiating classes."""
-
-    @overload
-    def load(self, cls: Type[obj]) -> obj:
-        pass
-
-    @overload
-    def load(self, cls: obj) -> obj:
-        pass
-
-    @abstractmethod
-    def load(self, cls: Type[obj] | obj) -> obj:
-        """Instantiates a class or returns an already instantiated instance."""
 
 
 class ClassInstantiator(ClassLoader):
